% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trmca.R
\name{.tmrca_onetree}
\alias{.tmrca_onetree}
\title{Calculate Time since Most Recent Common Ancestor for one tree}
\usage{
.tmrca_onetree(tree, states)
}
\arguments{
\item{states}{}
}
\description{
This function is for internal use only.
}
